Fixed the segfaults

+++ bonus/display.c
@@ -6,7 +6,7 @@
 /*   By: mialbert <mialbert@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/22 23:51:24 by mialbert          #+#    #+#             */
-/*   Updated: 2022/04/29 03:21:49 by mialbert         ###   ########.fr       */
+/*   Updated: 2022/05/02 14:22:19 by mialbert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	grey_screen(t_imgdata *data)
 {
-	size_t	x;
-	size_t	y;
+	int32_t	x;
+	int32_t	y;
 
 	x = 0;
 	y = 0;
